+++ src/onegov.town/HISTORY.rst
@@ -7,6 +7,9 @@ Unreleased
 - Adds the datetimepicker plugin.
   [msom]
 
+- Add retrieve password functionality.
+  [msom]
+
 0.2.6 (2015-07-16)
 ~~~~~~~~~~~~~~~~~~~